fucking fuck

+++ src/platform/server/bp-native/src/GameListener.cpp
@@ -31,7 +31,7 @@ int GameListener::popOldestEventId()
 {
     if (m_sLocalConsumedEventIds.size() > 0)
     {
-        short oldestEvent = m_sLocalConsumedEventIds.front();
+        int oldestEvent = m_sLocalConsumedEventIds.front();
 
         m_sLocalConsumedEventIds.erase(m_sLocalConsumedEventIds.begin());
